Moved the ColorPicker related classes and methods to a new class called. ComboBoxColorPicker.
